chg: check whether xray can start or not otherwise do not reset xray

+++ xray/run.sh
@@ -56,11 +56,13 @@ if [[ "$HTTP_PORTS" != "" ]];then
 	sed -i 's|"port":"80"|"port":"80,'$HTTP_PORTS'"|g' configs/05_inbounds_02_http_main.json
 fi
 
-rm -rf /dev/shm/hiddify-xtls-main.sock
-systemctl restart hiddify-xray.service
-systemctl start hiddify-xray.service
-systemctl status hiddify-xray.service
-
+xray run -test -confdir configs
+if  [[ $? == 0 ]];then
+	rm -rf /dev/shm/hiddify-xtls-main.sock
+	systemctl restart hiddify-xray.service
+	systemctl start hiddify-xray.service
+	systemctl status hiddify-xray.service
+fi
